Fixes bug with CMaker management on executale generation
Every time an executable was generated a new CMaker object was created. This made any change made to the class field (default init or user-made) be lost. Now, a copy of the field  CMaker is created for each variant and each variant can manage that as needed.

Additionally, this also makes the default initialization of the CMaker object take the user-defined include directories.